SOme refactorinh
