+++ scripts/fasttransformer/download_t5.sh
@@ -1,12 +1,10 @@
-# Arg1 - NumGpus
-# Arg2 - ModelName
-# Install dependencies for t5 transformations
+# Arg1 - Model path on HuggingFace
 cd FasterTransformer/build
 
 # Download model
 apt-get update
 apt-get install git-lfs
 git lfs install
-git lfs clone https://huggingface.co/$2
+git lfs clone https://huggingface.co/$1
 
 cd ../..
